+++ src/CartItem.js
@@ -1,0 +1,30 @@
+import React, { Component } from 'react';
+
+const USCurrencyFormat = new Intl.NumberFormat('en-US', {
+  style: 'currency',
+  currency: 'USD'
+});
+
+class CartItem extends Component {
+  render() {
+      const summary = Object.keys(this.props.selected).map((feature, idx) => {
+          const featureHash = feature + '-' + idx;
+          const selectedOption = this.props.selected[feature];
+      
+          return (
+            <div className="summary__option" key={featureHash}>
+              <div className="summary__option__label">{feature} </div>
+              <div className="summary__option__value">{selectedOption.name}</div>
+              <div className="summary__option__cost">
+                {USCurrencyFormat.format(selectedOption.cost)}
+              </div>
+            </div>
+          );
+        });
+      return (
+        summary
+      );
+  }
+}
+
+export default CartItem;
