--- src/Components/Cart-Item.js
@@ -1,1 +1,0 @@
-import React, { Component } from 'react';
