--- src/Components/Customize.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-class Customize extends React {
-
-    render() {
-        return(
-            <form className="main__form">
-                <h2>Customize your laptop</h2>
-                {features}
-          </form>
-        )
-    }
-
-
-}
