--- src/Components/Heading.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-function Heading() {
-
-    return (
-        <div className="App">
-            <header>
-                <h1>ELF Computing | Laptops</h1>
-            </header>
-        </div>
-    )
-}
-
-export default Heading
-
