+++ src/Heading.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+class Heading extends Component {
+  render() {
+    return (
+      <header>
+          <h1>ELF Computing | Laptops</h1>
+      </header>
+    )
+  }
+}
+
+export default Heading
+
